move IPC socket and logs under XDG_RUNTIME_DIR/halley
